changed db servers to postgres due to mysql limitations

+++ result-service/src/main/java/com/xzinoviou/academia/resultservice/domain/dto/ResultDto.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 /**
@@ -25,7 +26,7 @@ public class ResultDto {
 
     private Long courseId;
 
-    private Double grade;
+    private BigDecimal grade;
 
     private OffsetDateTime examDate;
