+++ result-service/src/main/java/com/xzinoviou/academia/resultservice/domain/request/ResultCreateRequest.java
@@ -4,6 +4,7 @@ import com.xzinoviou.academia.resultservice.domain.jpa.Semester;
 import lombok.Data;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 /**
@@ -19,7 +20,7 @@ public class ResultCreateRequest {
 
     private Long courseId;
 
-    private Double grade;
+    private BigDecimal grade;
 
     private OffsetDateTime examDate;
